backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
openai-whisper
nltk
ollama
requests
transformers
